backend/requirements.txt
Django==1.8.5
django-filter==0.11.0
djangorestframework==3.2.4
Markdown==2.6.2
